traffic_refresher/requirements.txt
azure-functions
python-dotenv
pandas
plotly
kaleido
azure-data-tables
azure-storage-blob
